US #83 Tasks 103 and 104
Added a div that contains within it a read only text field that is set
up to display debugging information upon compilation. Styled with the
rest of the page and hidden from view until the user presses a button in
the menu (currently shares button image with console, can easily be
changed later).