US#227 Task#256 Resizing Elements
Specifically targeted wider screen monitors (1920 and 2560+) as well as
making adjustments to the mobile landscape view so as to not have an
editor larger than the page and to the mobile portrait view to cut down
on blank space. Also made minor changes to a few of the button image
tags within index.html, thoroughly tested, no conflicts.